SocialMedia/requirements.txt
Django
pillow
django-mptt
